backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
black==24.4.0
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
distlib==0.3.8
distro==1.9.0
dulwich==0.21.7
elastic-transport==8.13.0
elasticsearch==8.13.1
elasticsearch-dsl==8.13.1
fastapi==0.110.1
fastjsonschema==2.19.1
filelock==3.13.4
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.4
keyring==24.3.1
Mako==1.3.3
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
openai==1.17.1
packaging==24.0
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
safetensors==0.4.3
sentry-sdk==2.1.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12
tokenizers==0.19.1
tomlkit==0.12.4
torch==2.3.0
tqdm==4.66.2
transformers==4.41.0
trove-classifiers==2024.4.10
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
xattr==1.1.0
zipp==3.18.1
